Updated hook logic
Refactored to use before/afterprint events;
Updated config;
Added new scripts;
Updated readme.
