+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 .idea
 node_modules
+lib
